chromiumide/src/test/unit/common/depot_tools.test.ts
// Copyright 2023 The ChromiumOS Authors
// Use of this source code is governed by a BSD-style license that can be
// found in the LICENSE file.

import 'jasmine';
import * as vscode from 'vscode';
import * as config from '../../../../shared/app/services/config';
import * as depotTools from '../../../common/depot_tools';
import * as testing from '../../testing';
import * as fakes from '../../testing/fakes';

describe('depot_tools', () => {
  const {vscodeEmitters, vscodeSpy} = testing.installVscodeDouble();
  testing.installFakeConfigs(vscodeSpy, vscodeEmitters);

  it('adjusts PATH based on settings', async () => {
    await config.paths.depotTools.update('/opt/custom_depot_tools');

    expect((await depotTools.envForDepotTools()).PATH).toEqual(
      jasmine.stringMatching('^/opt/custom_depot_tools:.*:.*/depot_tools')
    );
  });
  it('adjusts PATH if settings empty', async () => {
    await config.paths.depotTools.update('');

    expect((await depotTools.envForDepotTools()).PATH).toEqual(
      jasmine.stringMatching('^.*:.*/depot_tools')
    );
  });
});

describe('depot_tools not found', () => {
  const {vscodeEmitters, vscodeSpy} = testing.installVscodeDouble();
  testing.installFakeConfigs(vscodeSpy, vscodeEmitters);
  const {fakeExec} = testing.installFakeExec();
  fakes.installFakeDepotTools(fakeExec, false);

  it('opens a file dialog', async () => {
    const userDir = vscode.Uri.file('/usr/picked/sub');
    vscodeSpy.window.showOpenDialog.and.resolveTo([userDir]);

    expect((await depotTools.envForDepotTools()).PATH).toEqual(
      jasmine.stringMatching('^/usr/picked/sub:.*')
    );

    expect(vscodeSpy.window.showOpenDialog).toHaveBeenCalledTimes(1);
  });
});
